Extract handwritten interception code into a separate class
This makes evolution of the interception handling more uniform.
