+++ subprojects/core/src/main/java/org/gradle/internal/classpath/transforms/InstrumentingClassTransform.java
@@ -16,7 +16,7 @@
 
 package org.gradle.internal.classpath.transforms;
 
-import org.codehaus.groovy.runtime.ProcessGroovyMethods;
+import com.google.common.collect.ImmutableList;
 import org.codehaus.groovy.runtime.callsite.CallSiteArray;
 import org.codehaus.groovy.vmplugin.v8.IndyInterface;
 import org.gradle.api.file.RelativePath;
@@ -38,16 +38,13 @@ import org.objectweb.asm.MethodVisitor;
 import org.objectweb.asm.Type;
 import org.objectweb.asm.tree.MethodNode;
 
-import java.io.File;
 import java.lang.invoke.CallSite;
 import java.lang.invoke.MethodHandles;
 import java.lang.invoke.MethodType;
 import java.util.Collection;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.Properties;
 
 import static org.gradle.internal.classanalysis.AsmConstants.ASM_LEVEL;
 import static org.gradle.internal.classpath.transforms.CommonTypes.NO_EXCEPTIONS;
@@ -57,12 +54,9 @@ import static org.objectweb.asm.Opcodes.ACC_PRIVATE;
 import static org.objectweb.asm.Opcodes.ACC_STATIC;
 import static org.objectweb.asm.Opcodes.ACC_SYNTHETIC;
 import static org.objectweb.asm.Opcodes.H_INVOKESTATIC;
-import static org.objectweb.asm.Opcodes.INVOKESPECIAL;
 import static org.objectweb.asm.Opcodes.INVOKESTATIC;
-import static org.objectweb.asm.Opcodes.INVOKEVIRTUAL;
 import static org.objectweb.asm.Type.INT_TYPE;
 import static org.objectweb.asm.Type.getMethodDescriptor;
-import static org.objectweb.asm.Type.getObjectType;
 import static org.objectweb.asm.Type.getType;
 
 public class InstrumentingClassTransform implements ClassTransform {
@@ -72,96 +66,11 @@ public class InstrumentingClassTransform implements ClassTransform {
      */
     private static final int DECORATION_FORMAT = 36;
 
-    private static final Type SYSTEM_TYPE = getType(System.class);
-    private static final Type INTEGER_TYPE = getType(Integer.class);
     private static final Type INSTRUMENTED_TYPE = getType(Instrumented.class);
-    private static final Type LONG_TYPE = getType(Long.class);
-    private static final Type BOOLEAN_TYPE = getType(Boolean.class);
-    public static final Type PROPERTIES_TYPE = getType(Properties.class);
     private static final Type BYTECODE_INTERCEPTOR_FILTER_TYPE = Type.getType(BytecodeInterceptorFilter.class);
 
-    private static final String RETURN_STRING_FROM_STRING = getMethodDescriptor(STRING_TYPE, STRING_TYPE);
-    private static final String RETURN_STRING_FROM_STRING_STRING = getMethodDescriptor(STRING_TYPE, STRING_TYPE, STRING_TYPE);
-    private static final String RETURN_STRING_FROM_STRING_STRING_STRING = getMethodDescriptor(STRING_TYPE, STRING_TYPE, STRING_TYPE, STRING_TYPE);
-    private static final String RETURN_INTEGER_FROM_STRING = getMethodDescriptor(INTEGER_TYPE, STRING_TYPE);
-    private static final String RETURN_INTEGER_FROM_STRING_INT = getMethodDescriptor(INTEGER_TYPE, STRING_TYPE, Type.INT_TYPE);
-    private static final String RETURN_INTEGER_FROM_STRING_INTEGER = getMethodDescriptor(INTEGER_TYPE, STRING_TYPE, INTEGER_TYPE);
-    private static final String RETURN_INTEGER_FROM_STRING_STRING = getMethodDescriptor(INTEGER_TYPE, STRING_TYPE, STRING_TYPE);
-    private static final String RETURN_INTEGER_FROM_STRING_INT_STRING = getMethodDescriptor(INTEGER_TYPE, STRING_TYPE, Type.INT_TYPE, STRING_TYPE);
-    private static final String RETURN_INTEGER_FROM_STRING_INTEGER_STRING = getMethodDescriptor(INTEGER_TYPE, STRING_TYPE, INTEGER_TYPE, STRING_TYPE);
-    private static final String RETURN_LONG_FROM_STRING = getMethodDescriptor(LONG_TYPE, STRING_TYPE);
-    private static final String RETURN_LONG_FROM_STRING_PRIMITIVE_LONG = getMethodDescriptor(LONG_TYPE, STRING_TYPE, Type.LONG_TYPE);
-    private static final String RETURN_LONG_FROM_STRING_LONG = getMethodDescriptor(LONG_TYPE, STRING_TYPE, LONG_TYPE);
-    private static final String RETURN_LONG_FROM_STRING_STRING = getMethodDescriptor(LONG_TYPE, STRING_TYPE, STRING_TYPE);
-    private static final String RETURN_LONG_FROM_STRING_PRIMITIVE_LONG_STRING = getMethodDescriptor(LONG_TYPE, STRING_TYPE, Type.LONG_TYPE, STRING_TYPE);
-    private static final String RETURN_LONG_FROM_STRING_LONG_STRING = getMethodDescriptor(LONG_TYPE, STRING_TYPE, LONG_TYPE, STRING_TYPE);
-    private static final String RETURN_PRIMITIVE_BOOLEAN_FROM_STRING = getMethodDescriptor(Type.BOOLEAN_TYPE, STRING_TYPE);
-    private static final String RETURN_PRIMITIVE_BOOLEAN_FROM_STRING_STRING = getMethodDescriptor(Type.BOOLEAN_TYPE, STRING_TYPE, STRING_TYPE);
-    private static final String RETURN_PROPERTIES = getMethodDescriptor(PROPERTIES_TYPE);
-    private static final String RETURN_PROPERTIES_FROM_STRING = getMethodDescriptor(PROPERTIES_TYPE, STRING_TYPE);
-    private static final String RETURN_VOID_FROM_PROPERTIES = getMethodDescriptor(Type.VOID_TYPE, PROPERTIES_TYPE);
-    private static final String RETURN_VOID_FROM_PROPERTIES_STRING = getMethodDescriptor(Type.VOID_TYPE, PROPERTIES_TYPE, STRING_TYPE);
     private static final String RETURN_CALL_SITE_ARRAY = getMethodDescriptor(getType(CallSiteArray.class));
     private static final String RETURN_VOID_FROM_CALL_SITE_ARRAY_BYTECODE_INTERCEPTOR = getMethodDescriptor(Type.VOID_TYPE, getType(CallSiteArray.class), BYTECODE_INTERCEPTOR_FILTER_TYPE);
-    private static final String RETURN_MAP = getMethodDescriptor(getType(Map.class));
-    private static final String RETURN_MAP_FROM_STRING = getMethodDescriptor(getType(Map.class), STRING_TYPE);
-
-    private static final Type PROCESS_TYPE = getType(Process.class);
-    private static final Type PROCESS_BUILDER_TYPE = getType(ProcessBuilder.class);
-    private static final Type RUNTIME_TYPE = getType(Runtime.class);
-    private static final Type PROCESS_GROOVY_METHODS_TYPE = getType(ProcessGroovyMethods.class);
-    private static final Type STRING_ARRAY_TYPE = getType(String[].class);
-    private static final Type FILE_TYPE = getType(File.class);
-    private static final Type LIST_TYPE = getType(List.class);
-
-    // ProcessBuilder().start() -> start(ProcessBuilder, String)
-    private static final String RETURN_PROCESS = getMethodDescriptor(PROCESS_TYPE);
-    private static final String RETURN_PROCESS_FROM_PROCESS_BUILDER_STRING = getMethodDescriptor(PROCESS_TYPE, PROCESS_BUILDER_TYPE, STRING_TYPE);
-    // ProcessBuilder.startPipeline(List) -> startPipeline(List, String)
-    private static final String RETURN_LIST_FROM_LIST = getMethodDescriptor(LIST_TYPE, LIST_TYPE);
-    private static final String RETURN_LIST_FROM_LIST_STRING = getMethodDescriptor(LIST_TYPE, LIST_TYPE, STRING_TYPE);
-
-    // Runtime().exec(String) -> exec(Runtime, String, String)
-    // ProcessGroovyMethods.execute(String) -> execute(String, String)
-    private static final String RETURN_PROCESS_FROM_STRING = getMethodDescriptor(PROCESS_TYPE, STRING_TYPE);
-    private static final String RETURN_PROCESS_FROM_RUNTIME_STRING_STRING = getMethodDescriptor(PROCESS_TYPE, RUNTIME_TYPE, STRING_TYPE, STRING_TYPE);
-    private static final String RETURN_PROCESS_FROM_STRING_STRING = getMethodDescriptor(PROCESS_TYPE, STRING_TYPE, STRING_TYPE);
-    // Runtime().exec(String[]) -> exec(Runtime, String[], String)
-    // ProcessGroovyMethods.execute(String[]) -> execute(String[], String)
-    private static final String RETURN_PROCESS_FROM_STRING_ARRAY = getMethodDescriptor(PROCESS_TYPE, STRING_ARRAY_TYPE);
-    private static final String RETURN_PROCESS_FROM_RUNTIME_STRING_ARRAY_STRING = getMethodDescriptor(PROCESS_TYPE, RUNTIME_TYPE, STRING_ARRAY_TYPE, STRING_TYPE);
-    private static final String RETURN_PROCESS_FROM_STRING_ARRAY_STRING = getMethodDescriptor(PROCESS_TYPE, STRING_ARRAY_TYPE, STRING_TYPE);
-    // ProcessGroovyMethods.execute(List) -> execute(List, String)
-    private static final String RETURN_PROCESS_FROM_LIST = getMethodDescriptor(PROCESS_TYPE, LIST_TYPE);
-    private static final String RETURN_PROCESS_FROM_LIST_STRING = getMethodDescriptor(PROCESS_TYPE, LIST_TYPE, STRING_TYPE);
-    // Runtime().exec(String, String[]) -> exec(Runtume, String, String[], String)
-    private static final String RETURN_PROCESS_FROM_STRING_STRING_ARRAY = getMethodDescriptor(PROCESS_TYPE, STRING_TYPE, STRING_ARRAY_TYPE);
-    private static final String RETURN_PROCESS_FROM_RUNTIME_STRING_STRING_ARRAY_STRING = getMethodDescriptor(PROCESS_TYPE, RUNTIME_TYPE, STRING_TYPE, STRING_ARRAY_TYPE, STRING_TYPE);
-    // Runtime().exec(String[], String[]) -> exec(Runtume, String[], String[], String)
-    private static final String RETURN_PROCESS_FROM_STRING_ARRAY_STRING_ARRAY = getMethodDescriptor(PROCESS_TYPE, STRING_ARRAY_TYPE, STRING_ARRAY_TYPE);
-    private static final String RETURN_PROCESS_FROM_RUNTIME_STRING_ARRAY_STRING_ARRAY_STRING = getMethodDescriptor(PROCESS_TYPE, RUNTIME_TYPE, STRING_ARRAY_TYPE, STRING_ARRAY_TYPE, STRING_TYPE);
-    // Runtime().exec(String, String[], File) -> exec(Runtime, String, String[], File, String)
-    // ProcessGroovyMethods.execute(String, String[], File) -> execute(String, String[], File, String)
-    private static final String RETURN_PROCESS_FROM_STRING_STRING_ARRAY_FILE = getMethodDescriptor(PROCESS_TYPE, STRING_TYPE, STRING_ARRAY_TYPE, FILE_TYPE);
-    private static final String RETURN_PROCESS_FROM_RUNTIME_STRING_STRING_ARRAY_FILE_STRING = getMethodDescriptor(PROCESS_TYPE, RUNTIME_TYPE, STRING_TYPE, STRING_ARRAY_TYPE, FILE_TYPE, STRING_TYPE);
-    private static final String RETURN_PROCESS_FROM_STRING_STRING_ARRAY_FILE_STRING = getMethodDescriptor(PROCESS_TYPE, STRING_TYPE, STRING_ARRAY_TYPE, FILE_TYPE, STRING_TYPE);
-    // Runtime().exec(String[], String[], File) -> exec(Runtime, String[], String[], File, String)
-    // ProcessGroovyMethods.execute(String[], String[], File) -> execute(String[], String[], File, String)
-    private static final String RETURN_PROCESS_FROM_STRING_ARRAY_STRING_ARRAY_FILE = getMethodDescriptor(PROCESS_TYPE, STRING_ARRAY_TYPE, STRING_ARRAY_TYPE, FILE_TYPE);
-    private static final String RETURN_PROCESS_FROM_RUNTIME_STRING_ARRAY_STRING_ARRAY_FILE_STRING = getMethodDescriptor(PROCESS_TYPE, RUNTIME_TYPE, STRING_ARRAY_TYPE, STRING_ARRAY_TYPE, FILE_TYPE, STRING_TYPE);
-    private static final String RETURN_PROCESS_FROM_STRING_ARRAY_STRING_ARRAY_FILE_STRING = getMethodDescriptor(PROCESS_TYPE, STRING_ARRAY_TYPE, STRING_ARRAY_TYPE, FILE_TYPE, STRING_TYPE);
-    // ProcessGroovyMethods.execute(List, String[], File) -> execute(List, String[], File, String)
-    private static final String RETURN_PROCESS_FROM_LIST_STRING_ARRAY_FILE = getMethodDescriptor(PROCESS_TYPE, LIST_TYPE, STRING_ARRAY_TYPE, FILE_TYPE);
-    private static final String RETURN_PROCESS_FROM_LIST_STRING_ARRAY_FILE_STRING = getMethodDescriptor(PROCESS_TYPE, LIST_TYPE, STRING_ARRAY_TYPE, FILE_TYPE, STRING_TYPE);
-    // ProcessGroovyMethods.execute(String, List, File) -> execute(String, List, File, String)
-    private static final String RETURN_PROCESS_FROM_STRING_LIST_FILE = getMethodDescriptor(PROCESS_TYPE, STRING_TYPE, LIST_TYPE, FILE_TYPE);
-    private static final String RETURN_PROCESS_FROM_STRING_LIST_FILE_STRING = getMethodDescriptor(PROCESS_TYPE, STRING_TYPE, LIST_TYPE, FILE_TYPE, STRING_TYPE);
-    // ProcessGroovyMethods.execute(String[], List, File) -> execute(String[], List, File, String)
-    private static final String RETURN_PROCESS_FROM_STRING_ARRAY_LIST_FILE = getMethodDescriptor(PROCESS_TYPE, STRING_ARRAY_TYPE, LIST_TYPE, FILE_TYPE);
-    private static final String RETURN_PROCESS_FROM_STRING_ARRAY_LIST_FILE_STRING = getMethodDescriptor(PROCESS_TYPE, STRING_ARRAY_TYPE, LIST_TYPE, FILE_TYPE, STRING_TYPE);
-    // ProcessGroovyMethods.execute(List, List, File) -> execute(List, List, File, String)
-    private static final String RETURN_PROCESS_FROM_LIST_LIST_FILE = getMethodDescriptor(PROCESS_TYPE, LIST_TYPE, LIST_TYPE, FILE_TYPE);
-    private static final String RETURN_PROCESS_FROM_LIST_LIST_FILE_STRING = getMethodDescriptor(PROCESS_TYPE, LIST_TYPE, LIST_TYPE, FILE_TYPE, STRING_TYPE);
 
     private static final String GROOVY_INDY_INTERFACE_TYPE = getType(IndyInterface.class).getInternalName();
 
@@ -227,7 +136,7 @@ public class InstrumentingClassTransform implements ClassTransform {
             MethodVisitor methodVisitor = super.visitMethod(access, name, descriptor, signature, exceptions);
             Lazy<MethodNode> asMethodNode = Lazy.unsafe().of(() -> {
                 Optional<MethodNode> methodNode = classData.readClassAsNode().methods.stream().filter(method ->
-                        Objects.equals(method.name, name) && Objects.equals(method.desc, descriptor) && Objects.equals(method.signature, signature)
+                    Objects.equals(method.name, name) && Objects.equals(method.desc, descriptor) && Objects.equals(method.signature, signature)
                 ).findFirst();
                 return methodNode.orElseThrow(() -> new IllegalStateException("could not find method " + name + " with descriptor " + descriptor));
             });
@@ -260,19 +169,24 @@ public class InstrumentingClassTransform implements ClassTransform {
     }
 
     private static class InstrumentingMethodVisitor extends MethodVisitorScope {
+        private static final AdhocInterceptors ADHOC_INTERCEPTORS = new AdhocInterceptors();
+
         private final InstrumentingVisitor owner;
         private final String className;
         private final Lazy<MethodNode> asNode;
-        private final Collection<JvmBytecodeCallInterceptor> externalInterceptors;
+        private final Collection<JvmBytecodeCallInterceptor> interceptors;
         private final BytecodeInterceptorFilter interceptorFilter;
 
-        public InstrumentingMethodVisitor(InstrumentingVisitor owner, MethodVisitor methodVisitor, Lazy<MethodNode> asNode, ClassData classData, JvmBytecodeInterceptorSet externalInterceptors) {
+        public InstrumentingMethodVisitor(InstrumentingVisitor owner, MethodVisitor methodVisitor, Lazy<MethodNode> asNode, ClassData classData, JvmBytecodeInterceptorSet interceptors) {
             super(methodVisitor);
             this.owner = owner;
             this.className = owner.className;
             this.asNode = asNode;
-            this.interceptorFilter = externalInterceptors.getOriginalFilter();
-            this.externalInterceptors = externalInterceptors.getInterceptors(classData);
+            this.interceptorFilter = interceptors.getOriginalFilter();
+            List<JvmBytecodeCallInterceptor> providedInterceptors = interceptors.getInterceptors(classData);
+            this.interceptors = interceptorFilter.matches(ADHOC_INTERCEPTORS)
+                ? ImmutableList.<JvmBytecodeCallInterceptor>builderWithExpectedSize(providedInterceptors.size() + 1).add(ADHOC_INTERCEPTORS).addAll(providedInterceptors).build()
+                : providedInterceptors;
         }
 
         @Override
@@ -280,15 +194,9 @@ public class InstrumentingClassTransform implements ClassTransform {
             if (opcode == INVOKESTATIC && visitINVOKESTATIC(owner, name, descriptor)) {
                 return;
             }
-            if (opcode == INVOKEVIRTUAL && visitINVOKEVIRTUAL(owner, name, descriptor)) {
-                return;
-            }
-            if (opcode == INVOKESPECIAL && visitINVOKESPECIAL(owner, name, descriptor)) {
-                return;
-            }
 
-            for (JvmBytecodeCallInterceptor generatedInterceptor : externalInterceptors) {
-                if (generatedInterceptor.visitMethodInsn(this, className, opcode, owner, name, descriptor, isInterface, asNode)) {
+            for (JvmBytecodeCallInterceptor interceptor : interceptors) {
+                if (interceptor.visitMethodInsn(this, className, opcode, owner, name, descriptor, isInterface, asNode)) {
                     return;
                 }
             }
@@ -296,177 +204,10 @@ public class InstrumentingClassTransform implements ClassTransform {
         }
 
         private boolean visitINVOKESTATIC(String owner, String name, String descriptor) {
-            // TODO - load the class literal instead of class name to pass to the methods on Instrumented
-            if (owner.equals(SYSTEM_TYPE.getInternalName())) {
-                if (name.equals("getProperty")) {
-                    if (descriptor.equals(RETURN_STRING_FROM_STRING)) {
-                        _LDC(binaryClassNameOf(className));
-                        _INVOKESTATIC(INSTRUMENTED_TYPE, "systemProperty", RETURN_STRING_FROM_STRING_STRING);
-                        return true;
-                    }
-                    if (descriptor.equals(RETURN_STRING_FROM_STRING_STRING)) {
-                        _LDC(binaryClassNameOf(className));
-                        _INVOKESTATIC(INSTRUMENTED_TYPE, "systemProperty", RETURN_STRING_FROM_STRING_STRING_STRING);
-                        return true;
-                    }
-                } else if (name.equals("getProperties") && descriptor.equals(RETURN_PROPERTIES)) {
-                    _LDC(binaryClassNameOf(className));
-                    _INVOKESTATIC(INSTRUMENTED_TYPE, "systemProperties", RETURN_PROPERTIES_FROM_STRING);
-                    return true;
-                } else if (name.equals("setProperties") && descriptor.equals(RETURN_VOID_FROM_PROPERTIES)) {
-                    _LDC(binaryClassNameOf(className));
-                    _INVOKESTATIC(INSTRUMENTED_TYPE, "setSystemProperties", RETURN_VOID_FROM_PROPERTIES_STRING);
-                    return true;
-                } else if (name.equals("setProperty") && descriptor.equals(RETURN_STRING_FROM_STRING_STRING)) {
-                    _LDC(binaryClassNameOf(className));
-                    _INVOKESTATIC(INSTRUMENTED_TYPE, "setSystemProperty", RETURN_STRING_FROM_STRING_STRING_STRING);
-                    return true;
-                } else if (name.equals("clearProperty") && descriptor.equals(RETURN_STRING_FROM_STRING)) {
-                    _LDC(binaryClassNameOf(className));
-                    _INVOKESTATIC(INSTRUMENTED_TYPE, "clearSystemProperty", RETURN_STRING_FROM_STRING_STRING);
-                    return true;
-                } else if (name.equals("getenv")) {
-                    if (descriptor.equals(RETURN_STRING_FROM_STRING)) {
-                        // System.getenv(String) -> String
-                        _LDC(binaryClassNameOf(className));
-                        _INVOKESTATIC(INSTRUMENTED_TYPE, "getenv", RETURN_STRING_FROM_STRING_STRING);
-                        return true;
-                    } else if (descriptor.equals(RETURN_MAP)) {
-                        // System.getenv() -> Map<String, String>
-                        _LDC(binaryClassNameOf(className));
-                        _INVOKESTATIC(INSTRUMENTED_TYPE, "getenv", RETURN_MAP_FROM_STRING);
-                        return true;
-                    }
-                }
-            } else if (owner.equals(INTEGER_TYPE.getInternalName()) && name.equals("getInteger")) {
-                if (descriptor.equals(RETURN_INTEGER_FROM_STRING)) {
-                    _LDC(binaryClassNameOf(className));
-                    _INVOKESTATIC(INSTRUMENTED_TYPE, "getInteger", RETURN_INTEGER_FROM_STRING_STRING);
-                    return true;
-                }
-                if (descriptor.equals(RETURN_INTEGER_FROM_STRING_INT)) {
-                    _LDC(binaryClassNameOf(className));
-                    _INVOKESTATIC(INSTRUMENTED_TYPE, "getInteger", RETURN_INTEGER_FROM_STRING_INT_STRING);
-                    return true;
-                }
-                if (descriptor.equals(RETURN_INTEGER_FROM_STRING_INTEGER)) {
-                    _LDC(binaryClassNameOf(className));
-                    _INVOKESTATIC(INSTRUMENTED_TYPE, "getInteger", RETURN_INTEGER_FROM_STRING_INTEGER_STRING);
-                    return true;
-                }
-            } else if (owner.equals(LONG_TYPE.getInternalName()) && name.equals("getLong")) {
-                if (descriptor.equals(RETURN_LONG_FROM_STRING)) {
-                    _LDC(binaryClassNameOf(className));
-                    _INVOKESTATIC(INSTRUMENTED_TYPE, "getLong", RETURN_LONG_FROM_STRING_STRING);
-                    return true;
-                }
-                if (descriptor.equals(RETURN_LONG_FROM_STRING_PRIMITIVE_LONG)) {
-                    _LDC(binaryClassNameOf(className));
-                    _INVOKESTATIC(INSTRUMENTED_TYPE, "getLong", RETURN_LONG_FROM_STRING_PRIMITIVE_LONG_STRING);
-                    return true;
-                }
-                if (descriptor.equals(RETURN_LONG_FROM_STRING_LONG)) {
-                    _LDC(binaryClassNameOf(className));
-                    _INVOKESTATIC(INSTRUMENTED_TYPE, "getLong", RETURN_LONG_FROM_STRING_LONG_STRING);
-                    return true;
-                }
-            } else if (owner.equals(BOOLEAN_TYPE.getInternalName()) && name.equals("getBoolean") && descriptor.equals(RETURN_PRIMITIVE_BOOLEAN_FROM_STRING)) {
-                _LDC(binaryClassNameOf(className));
-                _INVOKESTATIC(INSTRUMENTED_TYPE, "getBoolean", RETURN_PRIMITIVE_BOOLEAN_FROM_STRING_STRING);
-                return true;
-            } else if (owner.equals(PROCESS_GROOVY_METHODS_TYPE.getInternalName()) && name.equals("execute")) {
-                Optional<String> instrumentedDescriptor = getInstrumentedDescriptorForProcessGroovyMethodsExecuteDescriptor(descriptor);
-                if (!instrumentedDescriptor.isPresent()) {
-                    return false;
-                }
-                _LDC(binaryClassNameOf(className));
-                _INVOKESTATIC(INSTRUMENTED_TYPE, "execute", instrumentedDescriptor.get());
-                return true;
-            }
-            if (owner.equals(PROCESS_BUILDER_TYPE.getInternalName()) && name.equals("startPipeline") && descriptor.equals(RETURN_LIST_FROM_LIST)) {
-                _LDC(binaryClassNameOf(className));
-                _INVOKESTATIC(INSTRUMENTED_TYPE, "startPipeline", RETURN_LIST_FROM_LIST_STRING);
-                return true;
-            } else if (owner.equals(className) && name.equals(CREATE_CALL_SITE_ARRAY_METHOD) && descriptor.equals(RETURN_CALL_SITE_ARRAY)) {
+            if (owner.equals(className) && name.equals(CREATE_CALL_SITE_ARRAY_METHOD) && descriptor.equals(RETURN_CALL_SITE_ARRAY)) {
                 _INVOKESTATIC(className, INSTRUMENTED_CALL_SITE_METHOD, RETURN_CALL_SITE_ARRAY);
                 return true;
             }
-            return false;
-        }
-
-        private boolean visitINVOKEVIRTUAL(String owner, String name, String descriptor) {
-            // Runtime.exec(...)
-            if (owner.equals(RUNTIME_TYPE.getInternalName()) && name.equals("exec")) {
-                Optional<String> instrumentedDescriptor = getInstrumentedDescriptorForRuntimeExecDescriptor(descriptor);
-                if (!instrumentedDescriptor.isPresent()) {
-                    return false;
-                }
-                _LDC(binaryClassNameOf(className));
-                _INVOKESTATIC(INSTRUMENTED_TYPE, "exec", instrumentedDescriptor.get());
-                return true;
-            }
-            if (owner.equals(PROCESS_BUILDER_TYPE.getInternalName())) {
-                if (name.equals("start") && descriptor.equals(RETURN_PROCESS)) {
-                    _LDC(binaryClassNameOf(className));
-                    _INVOKESTATIC(INSTRUMENTED_TYPE, "start", RETURN_PROCESS_FROM_PROCESS_BUILDER_STRING);
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        private Optional<String> getInstrumentedDescriptorForProcessGroovyMethodsExecuteDescriptor(String descriptor) {
-            if (descriptor.equals(RETURN_PROCESS_FROM_STRING)) {
-                // execute(String)
-                return Optional.of(RETURN_PROCESS_FROM_STRING_STRING);
-            } else if (descriptor.equals(RETURN_PROCESS_FROM_STRING_ARRAY)) {
-                // execute(String[])
-                return Optional.of(RETURN_PROCESS_FROM_STRING_ARRAY_STRING);
-            } else if (descriptor.equals(RETURN_PROCESS_FROM_LIST)) {
-                // execute(List)
-                return Optional.of(RETURN_PROCESS_FROM_LIST_STRING);
-            } else if (descriptor.equals(RETURN_PROCESS_FROM_STRING_STRING_ARRAY_FILE)) {
-                // execute(String, String[], File)
-                return Optional.of(RETURN_PROCESS_FROM_STRING_STRING_ARRAY_FILE_STRING);
-            } else if (descriptor.equals(RETURN_PROCESS_FROM_STRING_ARRAY_STRING_ARRAY_FILE)) {
-                // execute(String[], String[], File)
-                return Optional.of(RETURN_PROCESS_FROM_STRING_ARRAY_STRING_ARRAY_FILE_STRING);
-            } else if (descriptor.equals(RETURN_PROCESS_FROM_LIST_STRING_ARRAY_FILE)) {
-                // execute(List, String[], File)
-                return Optional.of(RETURN_PROCESS_FROM_LIST_STRING_ARRAY_FILE_STRING);
-            } else if (descriptor.equals(RETURN_PROCESS_FROM_STRING_LIST_FILE)) {
-                // execute(String, List, File)
-                return Optional.of(RETURN_PROCESS_FROM_STRING_LIST_FILE_STRING);
-            } else if (descriptor.equals(RETURN_PROCESS_FROM_STRING_ARRAY_LIST_FILE)) {
-                // execute(String[], List, File)
-                return Optional.of(RETURN_PROCESS_FROM_STRING_ARRAY_LIST_FILE_STRING);
-            } else if (descriptor.equals(RETURN_PROCESS_FROM_LIST_LIST_FILE)) {
-                // execute(List, List, File)
-                return Optional.of(RETURN_PROCESS_FROM_LIST_LIST_FILE_STRING);
-            }
-            // It is some signature of ProcessGroovyMethods.execute that we don't know about.
-            return Optional.empty();
-        }
-
-        private Optional<String> getInstrumentedDescriptorForRuntimeExecDescriptor(String descriptor) {
-            if (descriptor.equals(RETURN_PROCESS_FROM_STRING)) {
-                return Optional.of(RETURN_PROCESS_FROM_RUNTIME_STRING_STRING);
-            } else if (descriptor.equals(RETURN_PROCESS_FROM_STRING_ARRAY)) {
-                return Optional.of(RETURN_PROCESS_FROM_RUNTIME_STRING_ARRAY_STRING);
-            } else if (descriptor.equals(RETURN_PROCESS_FROM_STRING_STRING_ARRAY)) {
-                return Optional.of(RETURN_PROCESS_FROM_RUNTIME_STRING_STRING_ARRAY_STRING);
-            } else if (descriptor.equals(RETURN_PROCESS_FROM_STRING_ARRAY_STRING_ARRAY)) {
-                return Optional.of(RETURN_PROCESS_FROM_RUNTIME_STRING_ARRAY_STRING_ARRAY_STRING);
-            } else if (descriptor.equals(RETURN_PROCESS_FROM_STRING_STRING_ARRAY_FILE)) {
-                return Optional.of(RETURN_PROCESS_FROM_RUNTIME_STRING_STRING_ARRAY_FILE_STRING);
-            } else if (descriptor.equals(RETURN_PROCESS_FROM_STRING_ARRAY_STRING_ARRAY_FILE)) {
-                return Optional.of(RETURN_PROCESS_FROM_RUNTIME_STRING_ARRAY_STRING_ARRAY_FILE_STRING);
-            }
-            // It is some signature of Runtime.exec that we don't know about.
-            return Optional.empty();
-        }
-
-        private boolean visitINVOKESPECIAL(String owner, String name, String descriptor) {
             return false;
         }
 
@@ -495,10 +236,6 @@ public class InstrumentingClassTransform implements ClassTransform {
                 default:
                     throw new UnsupportedOperationException("Unknown interceptor request: " + interceptorFilter);
             }
-        }
-
-        private String binaryClassNameOf(String className) {
-            return getObjectType(className).getClassName();
         }
 
         private boolean isGroovyIndyCallsite(Handle bootstrapMethodHandle) {
